+++ src/frontend/apps/standalone_site/src/features/Contents/features/featureLoader.tsx
@@ -6,7 +6,9 @@
  *  - featureShuffles: Load the shuffles of the children features, used in the frontend page @see ContentsShuffle
  */
 
-import { useContentFeatures } from '../store/contentsStore';
+import { Fragment } from 'react';
+
+import { FeatureSample, useContentFeatures } from '../store/contentsStore';
 
 import {
   classRoomContents,
@@ -17,20 +19,61 @@ import {
 import { liveContents, LiveRouter, routesLive } from './Live';
 import { videoContents, VideoRouter, routesVideo } from './Video';
 
-const featureLoader = () => {
+enum RESOURCES_CHOICES {
+  VIDEO = 'video',
+  WEBINAR = 'webinar',
+  CLASSROOM = 'classroom',
+}
+
+const featureLoader = (inactive_content_types: string[]) => {
+  const featureRouter = [];
+  let featureRoutes = {};
+  const featureSamples: ((playlistId?: string) => FeatureSample)[] = [];
+  const featureShuffles = [];
+  if (!inactive_content_types.includes(RESOURCES_CHOICES.VIDEO)) {
+    featureRouter.push(
+      <Fragment key="videoRouter">
+        <VideoRouter />
+      </Fragment>,
+    );
+    featureSamples.push(videoContents);
+    featureRoutes = { ...featureRoutes, ...routesVideo };
+  }
+
+  if (!inactive_content_types.includes(RESOURCES_CHOICES.WEBINAR)) {
+    featureRouter.push(
+      <Fragment key="liveRouter">
+        <LiveRouter />
+      </Fragment>,
+    );
+    featureSamples.push(liveContents);
+    featureRoutes = { ...featureRoutes, ...routesLive };
+  }
+
+  if (!inactive_content_types.includes(RESOURCES_CHOICES.CLASSROOM)) {
+    featureRouter.push(
+      <Fragment key="classRoomRouter">
+        <ClassRoomRouter />
+      </Fragment>,
+    );
+    featureSamples.push(classRoomContents);
+    featureRoutes = { ...featureRoutes, ...routesClassRoom };
+    featureShuffles.push(
+      <Fragment key="classRoomShuffle">
+        <ClassRoomShuffle />
+      </Fragment>,
+    );
+  }
+
   useContentFeatures.setState({
-    featureRouter: [
-      <VideoRouter key="videoRouter" />,
-      <ClassRoomRouter key="classRoomRouter" />,
-      <LiveRouter key="liveRouter" />,
-    ],
-    featureRoutes: { ...routesVideo, ...routesLive, ...routesClassRoom },
-    featureSamples: (playlistId) => [
-      videoContents(playlistId),
-      liveContents(playlistId),
-      classRoomContents(playlistId),
-    ],
-    featureShuffles: [<ClassRoomShuffle key="classRoomShuffle" />],
+    featureRouter,
+    featureRoutes,
+    featureSamples: (playlistId) => {
+      return featureSamples.map((sample) => {
+        return sample(playlistId);
+      });
+    },
+    featureShuffles,
   });
 };