+++ src/frontend/apps/standalone_site/src/features/Playlist/features/UpdatePlaylist/components/UpdatePlaylistPage.spec.tsx
@@ -132,7 +132,7 @@ describe('<UpdatePlaylistPage />', () => {
   });
 
   it('checks the contents render by playlist', async () => {
-    featureContentLoader();
+    featureContentLoader([]);
 
     render(<UpdatePlaylistPage />);